Update FAQ with section on Rawhide vs ELN kernels
While the existing section on Rawhide vs ELN does mention that listing the differences is beyond the scope of the document, it is important to lay out some of the differences between the Rawhide and ELN kernels, as there are many supported Fedora configurations that simply will not run with the ELN kernel.  I have called out btrfs specifically as it is the default file system for a number of Fedora editions, and I have already seen a few reports of users trying to install the ELN kernel on Fedora systems with btrfs who were confused as to why it did not work.